third_party/WebKit/LayoutTests/http/tests/serviceworker/activation-after-registration.html

<!DOCTYPE html>
<title>Service Worker: Activation occurs after registration</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<body>
<script>
var t = async_test('activation occurs after registration');
t.step(function() {
    var scope = 'resources/blank.html';
    var registration;

    service_worker_unregister_and_register(
        t, 'resources/empty-worker.js', scope)
      .then(t.step_func(function(r) {
          registration = r;
          return wait_for_update(t, registration);
        }))
      .then(t.step_func(function(worker) {
          assert_equals(worker.state, 'parsed', 'worker should be in the "parsed" state upon registration');
          worker.addEventListener('statechange', t.step_func(function(event) {
              if (event.target.state == 'activated')
                  service_worker_unregister_and_done(t, scope);
          }));
        }))
      .catch(unreached_rejection(t));
});
</script>
</body>
